AI/AMIT WORK DSU/translation_requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai-whisper==20231117
deep-translator==1.11.4
numpy==1.24.3
torch==2.0.1
python-dotenv==1.0.0
soundfile==0.12.1
librosa==0.10.1
scipy==1.11.4
